Update the_beginning.snbt
Unable to complete these two Genesis quests playing peaceful:

Lost Cities
requires skulls. would be nice if we could buy them via monicoin as these are not lootable in Peaceful. Else recommend moving to The Beginning near Simulating Mobs, but a floating quest not dependent on it.

Electromagnet
requires Vibrant Crystal, which requires an Autoclave, Extractor, and Alloy Smelter. These machines are not within the Genesis tree. Suggest allowing for monicoin purchase of Vibrant Crystal as these are not lootable in Peaceful else recommend moving to The Beginning dependent on creating an Autoclave.

Inspector Gadget
requires Ender Pearls, which if one is not looting, can only be made if one has an Alloy Smelter. Recommend if not monicoin for Ender Pearls then moving this to The Beginning, dependent on Pulsating Iron and Ender Pearls.