Alternate textures for Omnic Matrix casings
This is an edit based on HSS-E casings, I think the current texture doesn't look vibrant enough and the diamond colored patterns clash too hard with the rest of the texture and add too much visual noise. Instead, the connected textures reveal a woven lattice pattern reminiscent of the diamond lattices that go into making the crystal matrix ingots needed for this casing.