Alternate texture for Omnic Matrix coils
This is a texture I made before omnic coils were updated, with some last minute edits. I went for a more subtle diamond color for the frame to help it clash less with the rest of the multiblock, used very bright colors for the active texture (this thing is supposed to burn hot!) and went with darker, desaturated colors for the inactive texture to make the active texture stand out even more when the machine turns on.