Various texture updates (#1205)
* Modernized textures for Singularity Containment Unit and Contained Singularity

* Modernized texture for Field Stabilized Omnic Pulsar Compound

* Added some basic shading to Dimensional Circuit Board and moved Dimensional Circuit Board & Processing Unit down by 1px to match other circuit boards

* Modernized texture for Magnetron

* Modernized textures for Netherite line items

* Modernized textures for Holmium line items

* Modernized textures for Microverse casings

* Alternate textures for Netherite casings

This is a new texture made from scratch, I used the same color palette as the texture for netherite blocks as I believe the current palette is way too dark and clashes pretty hard with the rest of the multiblock (both the superassembler and the hyperbolic projector).

* Alternate textures for Omnic Matrix casings

This is an edit based on HSS-E casings, I think the current texture doesn't look vibrant enough and the diamond colored patterns clash too hard with the rest of the texture and add too much visual noise. Instead, the connected textures reveal a woven lattice pattern reminiscent of the diamond lattices that go into making the crystal matrix ingots needed for this casing.

* Alternate texture for Omnic Matrix coils

This is a texture I made before omnic coils were updated, with some last minute edits. I went for a more subtle diamond color for the frame to help it clash less with the rest of the multiblock, used very bright colors for the active texture (this thing is supposed to burn hot!) and went with darker, desaturated colors for the inactive texture to make the active texture stand out even more when the machine turns on.

* Alternate textures for end game wafers and chips

Not much to say about this, I made those pretty early, shortly after the complex SMD components got their updated textures. I went for simpler, easier to read patterns for the Multidimensional CPU and Quantum SoC and chose more vibrant colors so they fit in better with the rest of the GT Refreshed textures.

* Alternate textures for Universe-doped boules and wafers

The first textures I made for my own playthrough. For the boule I went with the same smooth look as the Neutronium doped boule since it felt more fitting, and did some hue shifting on the darker shades to make the items look more vibrant.