Tweaks & fixes (#1331)
* Update GT Energy Converter tooltip
This is to clarify that no, you can't use it with conduits in EU->FE mode. Yes, it's in the questbook, but you clearly didn't read it.

* Add warning that GT robot arm, fluid regulator don't work with pipes

* Make UXPIC wafers comparable in difficulty to UHPIC
instead of being easier

* Add PECA to plastic plates tag

* Change circuit tag icons to universal circuits.
Thanks to Lue for showing how to implement this!

* Lower voltage tier of tributyl phosphate rare earth electrolysis
Apparently this recipe was getting neglected, so this should make it more useful.

* Fix Caesium Hydroxide stoichiometry
Thanks Onion!

* Change Signalum dust to a catalyst in Neocryolite processing

* Fix Infinity ME fluid cell to only exist in NM

* Remove Slime Block fabrication
Slimeball fabrication is sufficient. Slime Block fabrication renders it entirely superfluous due to giving the same resource in larger amounts.

* Change Terbium Salt into a GT material for chemical formula in tooltip
This also leaves the smaller dust sizes available if we want to make it a byproduct for other processes we choose to rework in the future

* Unhide pasteurized milk from JEI/EMI
Fix for #1293

* Fix tetraethyllead voiding sodium

* Remove superfluous Mud recipe in create compat
GTM already adds this recipe, so no need for a duplicate

* Fix recipe conflict with Create Rose Quartz and Fluix

* Remove ME, Dense ME conduit recipes in EM