Modified HNN script to add a PPM recipe with an Autoclave (#1886)
* Modified HNN script to add the PPM recipe using
PBI(polybenzimidazole) and 8 Pulsating dust.

* Nerfed the recipe by halving PPM yields,
multiplying PBI 9, and increasing the duration,
it's still better than the ol' alloy smelter
method but not as broken.

* Lint go brrr,
ESLint hates singular quotes so I replaced those
with double quotes.

* Made it faster
Yeah 50 seconds per recipe was a bit much... my bad.

* Switched it back to autoclave
I was experimenting with it being a chemical bath
recipe but decided against it and forgot to swap
it back.

---------

Co-authored-by: Jaantheman <janjaydenw730@gmail.com>