Minor questbook fixes and Major Changes (#2014)
* questbook fixing montage.

* Add Assline fluid ordering for EM

* questbook fixes 2: electric boogaloo
rearranges positions of quests and modifies text and dependiencies of quests

* OCD fix(neutronium panel quest was misalligned)

* Black Steel can now be made with coal and charcoal
if a dev wants I can revert this, btw this is NM specific

* Daily fact of apop's stupidity
I forgot to change the IDs and a letter e lol

* Revert "Daily fact of apop's stupidity"

This reverts commit 08d410e009432e2b40aea4b5841918842706eb2a.

* Revert "Black Steel can now be made with coal and charcoal"

This reverts commit 6f54b5993b4c09b70b05ed622d3a3f979cca0887.

* fixed some oopsies.

* fixing some oopsies 2:electric boogaloo

* Colored PPM quest correctly

* Remove the magnetron from the naq superfuel line
it just doesn't exist why keep it lol(maybe it'll exist later but for now t's non-existant so I removed it)

* Corrected magnetron fix to put it back in HM
turns out it wasn't cut it was just an HM thing, fixed properly.

* Fixes the supercomputer bug
WARNING: Models are now Self Aware by default! Do not merge #2014 unless you're ok with this change!!!

* Revert "Fixes the supercomputer bug"

This reverts commit ffdc9e91f8ba19f5b4f3485f34b920eada8a493a.

* Allows framed doors to be used in the cleanroom