Texture Updates: Round III (#2130)
* Recolor Cryolobus & Cryococcus

* Crystal Matrix

* Mob Guts

* Energy Cores p1

This would be easier out of the website, but that doesn't make it *hard* to do like this.

* Energy Cores p2

Indeed, while this needs more commits it also forces me to not put all, like, what, five wholly separate changes in one monolith "Texture Updates: Round III" commit.

* Energy Cores p3

This is sosoSO much easier then model files what was I THINKING for naqfuel,,,

* yeet thissss

* yeet this too

* this one needs a yeet too

* so does this one need a yeet

Okay so this is one of those changes that would be easier on the desktop app

* Fix broken CTM on P.R.I.S.M. Glass

I don't know how this one slipped past me trying to fix it *twice* but hopefully third time's the charm

* Secondary Colors

* Rename Microversium to Microversite

This is sooooo sooooooooo petty but like   WHY DOES IT END WITH -IUM IF ITS NOT A METALLIC ELEMENT

* Thrusters p1

* Thrusters p2

* Thrusters p3

* Thrusters p4

* Electrical Steel

Yes this one was already changed in the secondary colors one I changed it again

* Secondary Colors II

* Thrusters p5 // Solidified Elements p1

* Thrusters p6

* Thrusters p7

* Solidified Elements p2

* Solidified Elements p3

* Solidified Elements p4

* Thorium't

I really *should* be doing this for all the new colors we added to match NCN but the others are nice, y'know? they make em more distinct which is good

* Thrusters p8

* NCN Isotopes & Fuels

Not as big as it seems!

* Solidified & Stabilized p5

* Oops!

* Solidified & Stabilized p6

* Solidified & Stabilized p6

* Solidified & Stabilized p7

* Stabilized & Solidified p7.5

* ..

* Add files via upload

* ...

* .....

* going dark

* Various Tweaks

* moar thruster stuff

* it goes on

* and on and on